Change Flim Op names
Thanks to @JenuC for the suggestions!
